% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DL_camels.R
\name{download_camels}
\alias{download_camels}
\title{Download CAMELS data}
\usage{
download_camels(data_dir)
}
\arguments{
\item{data_dir}{NEW directory path that will be created by this function, the filepath leading up to the final folder should already exist, while the last folder should not yet exist and will be created for you}
}
\value{
NA
}
\description{
downloads and unzips CAMELS data from https://ral.ucar.edu/solutions/products/camels into the user's chosen directory
}
\details{
more specifically this is linked to the version 1.2 timeseries meterology, observed flow, metadata here:
https://ral.ucar.edu/sites/default/files/public/product-tool/camels-catchment-attributes-and-meteorology-for-large-sample-studies-dataset-downloads/basin_timeseries_v1p2_metForcing_obsFlow.zip

as well as the CAMELS attribute data v 2.0 from here:
https://ral.ucar.edu/sites/default/files/public/product-tool/camels-catchment-attributes-and-meteorology-for-large-sample-studies-dataset-downloads/camels_attributes_v2.0.zip
}
\examples{
\donttest{
data_dir <- "~/CAMELS_data"
download_camels(data_dir)
}

}
