% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAMELS.R
\name{extract_huc_data}
\alias{extract_huc_data}
\title{Extract data from CAMELS by attribute}
\usage{
extract_huc_data(basin_dir, attr_dir, huc8_names)
}
\arguments{
\item{attr_dir}{directory path for attributes data (e.g: "~/camels_attributes_v2.0")}

\item{huc8_names}{vector of huc8 watershed IDs}

\item{daymet_dir}{directory path to basin data should look something like: "~/basin_dataset_public_v1p2/" this directory pathway MUST end with the daymet/ folder
-- within this folder there should more directories (labeled '01', '02', '03' etc) representing different huc2 watersheds}
}
\value{
named list for each huc8 with 8 nested named lists, one for each attribut of interest
}
\description{
This function takes a list of watershed Huc 8s and extracts daily mean forcing dayment data as well as relevant climate, geology, hyrdology, area, soil, topography and vegetation data
}
\examples{
\dontrun{
ts_dir <- "~/CAMELS/basin_dataset_public_v1p2"
attr_dir <- "~/CAMELS/camels_attributes_v2.0"
huc8_names <- c("01013500", "08269000")

data <- extract_huc_data(basin_dir = ts_dir, attr_dir = attr_dir, huc8_names = huc8_names)

## list of hucs queried
> names(data)
[1] "mean_forcing_daymet" "usgs_streamflow"     "camels_clim"         "camels_geol"         "camels_hydro"
[6] "camels_name"         "camels_soil"         "camels_topo"         "camels_vege"

}

}
