% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_site_data.R
\docType{data}
\name{trial_sites}
\alias{trial_sites}
\title{list of trial sites for the get_sample_data() function}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 34 rows and 7 columns.
}
\usage{
trial_sites
}
\description{
list of trial sites for the get_sample_data() function
}
\keyword{datasets}
