% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USGS_parameter}
\alias{USGS_parameter}
\title{Parameter codes of interest for USGS water quality data}
\format{
A data frame with 30 rows and 4 variables:
\describe{
  \item{X5_digit_code}{usgs parameter code}
  \item{Parameter_USGS}{name + brief explination of code}
  \item{Parameter}{parameter name}
  \item{Unit}{units for parameter}
#' }
}
\usage{
USGS_parameter
}
\description{
List of codes and their meanings
}
\keyword{datasets}
